[PyTorch Mobile] Eliminate static default_extra_files_mobile from header import.h
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50795

There's [a post](https://fb.workplace.com/groups/2148543255442743/permalink/2583012411995823/) about a customer having to pass in `-Wno-global-constructors` to disable warnings related to calling constructors for global objects. This is related to the initialization of `default_extra_files_mobile` in `import.h`.

It requires end users to pass in the compiler flag, since the definition is now in code (.cpp files) that they will be compiling.

In addition, it makes the API for `_load_for_mobile` non-re-entrant (i.e. can not be safely used concurrently from multiple threads without the caller taking a mutex/lock) if the `extra_files_mobile` argument is not explicitly passed in.

Instead, a better option would be to create different overloads; one which requires all 3 parameters, and one that can work with 1-2. This solves the problem without creating a static variable.

ghstack-source-id: 120055794

Differential Revision: [D25968216](https://our.internmc.facebook.com/intern/diff/D25968216/)